Rename nvim-tree plugin spec
